Some fixs and added base xtoa

+++ srcs/ft_ltoa_base.c
@@ -6,7 +6,7 @@
 /*   By: acazuc <acazuc@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/02/05 11:20:57 by acazuc            #+#    #+#             */
-/*   Updated: 2016/02/05 11:24:21 by acazuc           ###   ########.fr       */
+/*   Updated: 2016/02/05 11:34:49 by acazuc           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 static size_t	get_size(long long int n, char *base)
 {
-	size_t	size;
+	size_t					size;
 
 	size = n < 0 ? 2 : 1;
 	n = n < 0 ? -n : n;
@@ -28,11 +28,11 @@ static size_t	get_size(long long int n, char *base)
 
 char			*ft_ltoa_base(long long int n, char *base)
 {
-	char	*result;
-	size_t	size;
-	size_t	j;
-	size_t	i;
-	size_t	nb;
+	char					*result;
+	size_t					size;
+	unsigned long long int	j;
+	unsigned long long int	i;
+	unsigned long long int	nb;
 
 	if (!base || ft_strlen(base) < 2)
 		return (NULL);
